Add property to toggle isalin on/off
