+++ src/main/java/com/aclt/isalin/properties/IsalinProperties.java
@@ -4,8 +4,10 @@ import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
 @Data
-@ConfigurationProperties(prefix = "aclt.isalin")
+@ConfigurationProperties(prefix = "isalin")
 public class IsalinProperties {
+
+    private boolean enabled = true;
 
     private String serviceUrl = "https://translation.googleapis.com/language/translate/v2";
