--- src/test/resources/application-test.properties
@@ -1,1 +1,0 @@
-aclt.isalin.service-key=<INSERT_KEY_HERE>
