+++ src/test/resources/isalin-disabled.properties
@@ -1,0 +1,1 @@
+isalin.enabled=false
