+++ src/test/resources/isalin-enabled.properties
@@ -1,0 +1,1 @@
+isalin.service-key=<INSERT_KEY_HERE>
